Extract controller code into reusable package
